feat: change KC_MEDIA_PLAY_PAUSE to C-F4
The function media play/pause does its work on any media, which means
that also YouTube panes are controlled - actually the last media opened.
that's not what I want. I want to control Music.app only. AS calling a
command or function is not possible, I trigger a complex modification in
Karabiner Elements that runs an Applescript command to talk to the Music.app.